finished time for test

+++ src/main/java/io/rtiga/onlinepajak/domains/dto/FakturData.java
@@ -4,6 +4,7 @@ package io.rtiga.onlinepajak.domains.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,10 +14,11 @@ public class FakturData {
     private String npwpPenjual;
     private String namaPembeli;
     private String npwpPembeli;
-    private Double hargaTotal;
-    private Double ppn;
-    private Double ppnbm;
+    private BigDecimal hargaTotal;
+    private BigDecimal ppn;
+    private BigDecimal ppnbm;
     private LocalDate tanggalFaktur;
     private String nomorFaktur;
     private String qrCodeValue;
+
 }
